Keep invited players in party lobby
