+++ src/Home.jsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from "react";
-import { HashRouter, Route, Routes } from "react-router-dom";
+import { HashRouter, Navigate, Route, Routes, useLocation } from "react-router-dom";
 
 /***
  * https://blog.logrocket.com/lazy-loading-components-in-react-16-6-6cea535c0b52/
@@ -25,6 +25,29 @@ import ErrorBoundary from "./component/ErrorBoundary/ErrorBoundary.jsx";
 
 import { formatWording } from "../utils/langUtils";
 
+const ROOM_CODE_PATTERN = /^\d{6}$/;
+
+const PartyRoute = () => {
+    const location = useLocation();
+    const inviteRoomCode = new URLSearchParams(location.search).get('room')?.trim() || '';
+
+    if (inviteRoomCode) {
+        return (
+            <Navigate
+                replace
+                to="/"
+                state={{
+                    stage: "party_setup",
+                    roomID: inviteRoomCode,
+                    notice: ROOM_CODE_PATTERN.test(inviteRoomCode) ? "" : formatWording("error.invalid.inputRoom", {}),
+                }}
+            />
+        );
+    }
+
+    return <PartyPage/>;
+};
+
 const Home = () => {
     return (
         <ErrorBoundary fallback={<p>{formatWording("error.load.page", {})}</p>}>
@@ -32,7 +55,7 @@ const Home = () => {
                 <HashRouter>
                     <Routes>
                         <Route path="/local" element={<MainPage/>}/>
-                        <Route path="/party" element={<PartyPage/>}/>
+                        <Route path="/party" element={<PartyRoute/>}/>
                         <Route path="/" element={<OpeningPageV2/>}/>
                     </Routes>
                 </HashRouter>
